backend/requirements.txt
opcua>=0.98.13
fastapi>=0.115.0
uvicorn>=0.30.0
mysql-connector-python>=8.0.0
cryptography>=40.0.0
# Ajoute d'autres packages si tu en utilises (ex: numpy, pandas, etc.)